a.k.a signalBasedManager
This network manager was developed to improve the establishment of the wireless network connection of mobile users in highly dynamic scenarios, due that the standard GNOME NetworkManager of GNU Linux is not suitable for these environments. 

For more information about its evaluation, you can find it on this open access article http://www.hindawi.com/journals/misy/2016/4015625/
please feel free to use and please don't forget to cite us as a reference in your work.